backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Pillow==10.2.0
youtube-transcript-api==0.6.2 
